Changed keybindings to trigger on key up
This is probably not the behavior we want, but having it on pointer down
meant the focused text input registerd the key as it was being released;
so it a bound key 'r' triggered the focus of a text input, the input box
got an 'r' in it. This removes that problem for now.